requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.1
django-ratelimit==3.0.1
django-session-timeout==0.1.0
djangorestframework==3.13.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
